♻️ refactored `aswe.core.use_case`
